Make all C API tests unit tests
This change integrates all end-to-end C API tests into the respective
modules in the form of unit tests. Doing so arguably makes more sense
for the capi crate, because
- there aren't really any visibility concerns for the C API and
- that means we no longer have to build as a `lib` crate type
