+++ capi/Cargo.toml
@@ -7,8 +7,7 @@ rust-version = "1.64"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
-# `lib` is necessary for end-to-end tests.
-crate-type = ["lib", "cdylib", "staticlib"]
+crate-type = ["cdylib", "staticlib"]
 
 [features]
 # Enable this feature to re-generate the library's C header file. An
